+++ core/config_manager.py
@@ -51,7 +51,6 @@ class ConfigManager:
             "yt_dlp_version": "Unknown",
             "cookies_from_browser": "None",
             "gallery_cookies_from_browser": "None",
-            "js_runtime_path": "",  # New setting for JavaScript runtime
             "output_template": "%(title)s [%(uploader)s][%(release_date>%m-%d-%Y)s][%(id)s].%(ext)s",
             "embed_chapters": "True",
             "subtitles_embed": "False",
@@ -86,7 +85,6 @@ class ConfigManager:
             "yt_dlp_version": "Unknown",
             "cookies_from_browser": "None",
             "gallery_cookies_from_browser": "None",
-            "js_runtime_path": "",  # New setting for JavaScript runtime
             "output_template": "%(title)s [%(uploader)s][%(release_date>%m-%d-%Y)s][%(id)s].%(ext)s",
             "embed_chapters": "True",
             "subtitles_embed": "False",